# rpc,rest,zmq: faster getblock, NotifyBlock and rest_block by reading raw block