@andrewtoth 2022-10-30
For some more context, here is the PR that introduced `ReadRawBlockFromDisk` https://github.com/bitcoin/bitcoin/pull/13151.