@andrewtoth 2022-10-30
@luke-jr I wasn't aware of #21319 until now, but I can't comment there because it's locked. If you want to reopen that I can review and remove the change to RPC in this PR so it will only be for REST.