@andrewtoth 2023-11-28
> Does testing the different branches involve something more than checking out the different branches (master and read-raw-block) on my local computer?

@kashifs after checkout out each branch, you must again compile from source, and shut down the running bitcoind and then run the newly compiled binary at ./src/bitcoind.