@kashifs 2023-11-30
@andrewtoth, okay thanks. I did as you suggested and was able to get similar speed improvements when testing the `rest` command. 10.260ms on branch `master` and 1.014ms on branch `bitcoin-read-raw-block`. I only see improvements when running the `rest` testing command more than once though. Is this the expected behavior? I hope it's not due to some caching or other improvements unrelated to your code updates. Also, I was not able to appreciate any improvements when running the `RPC` command. Could there be something else that I'm missing? Thanks again for your time.